check queue position directly on CartQueue
